Finished renaming device pointer variables including their comments. Also Foo.h
