+++ Simulation/Edges/AllDSSynapses.h
@@ -219,27 +219,28 @@ protected:
         *  Delete GPU memories.
         *  (Helper function of deleteSynapseDeviceStruct)
         *
-        *  @param  allSynapsesDevice  GPU address of the allSynapses struct on device memory.
+        *  @param  allSynapsesDeviceProps  GPU address of the allSynapses struct on device memory.
         */
-       void deleteDeviceStruct( AllDSSynapsesDeviceProperties& allSynapses );
+       void deleteDeviceStruct( AllDSSynapsesDeviceProperties& allSynapsesDeviceProps );
 
        /**
         *  Copy all synapses' data from host to device.
         *  (Helper function of copySynapseHostToDevice)
         *
-        *  @param  allSynapsesDevice     GPU address of the allSynapses struct on device memory.
-        *  @param  numNeurons            Number of neurons.
-        *  @param  maxSynapsesPerNeuron  Maximum number of synapses per neuron.
+        *  @param  allSynapsesDevice      GPU address of the allSynapses struct on device memory.
+        *  @param  allSynapsesDeviceProps GPU address of the AllDSSSynapses struct on device memory.
+        *  @param  numNeurons             Number of neurons.
+        *  @param  maxSynapsesPerNeuron   Maximum number of synapses per neuron.
         */
-       void copyHostToDevice( void* allSynapsesDevice, AllDSSynapsesDeviceProperties& allSynapses, int numNeurons, int maxSynapsesPerNeuron );
+       void copyHostToDevice( void* allSynapsesDevice, AllDSSynapsesDeviceProperties& allSynapsesDeviceProps, int numNeurons, int maxSynapsesPerNeuron );
 
        /**
         *  Copy all synapses' data from device to host.
         *  (Helper function of copySynapseDeviceToHost)
         *
-        *  @param  allSynapses  GPU address of the allSynapses struct on device memory.
+        *  @param  allSynapsesDeviceProps  GPU address of the allSynapses struct on device memory.
         */
-       void copyDeviceToHost( AllDSSynapsesDeviceProperties& allSynapses);
+       void copyDeviceToHost( AllDSSynapsesDeviceProperties& allSynapsesDeviceProps);
 #else // !defined(USE_GPU)
 protected:
    /**
