+++ Testing/Core/OperationManagerTestingClass.h
