+++ Testing/Core/OperationManagerTests.cpp
@@ -11,7 +11,7 @@
 
 #include "gtest/gtest.h"
 
-#include "Foo.h"
+#include "OperationManagerTestingClass.h"
 #include "Simulation/Core/OperationManager.h"
 #include "Simulation/Core/Operations.h"
 #include "Layout.h"
